refactor(Favorite*.*): 코드정리, 즐겨찾기 인수테스트 추가
- FavoriteContoller 코드정리(import 순서 변경)
- FavoriteService, FavoriteRepository 사용하지 않는 로직 제거
- 즐겨찾기 인수테스트 추가
