+++ src/main/java/nextstep/subway/favorite/domain/FavoriteRepository.java
@@ -1,12 +1,11 @@
 package nextstep.subway.favorite.domain;
 
-import nextstep.subway.member.domain.Member;
-import nextstep.subway.station.domain.Station;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import nextstep.subway.member.domain.Member;
+
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
-    Optional<Favorite> findByMemberAndSourceAndTarget(Member member, Station source, Station target);
     Optional<Favorite> findByMember(Member member);
 }
